+++ src/solvers/GodonovSolver2D_t.cpp
@@ -53,7 +53,7 @@ double GodonovSolver2D_t<FluxCalculator>::calculate_delta_t(double cfl, FVM::Ent
 template<typename FluxCalculator>
 void GodonovSolver2D_t<FluxCalculator>::timestep(double delta_t, FVM::Entities::Mesh2D_t& mesh) {
     #pragma omp parallel for schedule(guided)
-    for (size_t i = 0; i < mesh.n_cells_; ++i) {
+    for (long long i = 0; i < mesh.n_cells_; ++i) { // Because msvc openmp wants a signed index, so no size_t :(
         FVM::Entities::Cell_t& cell =  mesh.cells_[i];
 
         double F_1 = 0;
